services/pulse-ai-core/requirements.txt
opensearch-py
pandas
scikit-learn
hdbscan
openai
pydantic-settings
uvicorn[standard]
fastapi
sqlalchemy
psycopg2-binary
bertopic